Исправлены: схемы, модели, роуты

+++ app/models/module.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, BigInteger, String, Text, Integer
+from sqlalchemy import Column, Integer, String, Text, Integer
 from sqlalchemy.orm import relationship
 from app.db.base import Base
 
 
 class Module(Base):
     __tablename__ = "modules"
-    id = Column(BigInteger, primary_key=True)
+    id = Column(Integer, primary_key=True)
     title = Column(String, nullable=False)
     description = Column(Text)
     sort_order = Column(Integer, default=0)
