+++ app/models/quiz.py
@@ -1,12 +1,12 @@
-from sqlalchemy import Column, BigInteger, Integer, ForeignKey, Text, JSON, Boolean, TIMESTAMP, func
+from sqlalchemy import Column, Integer, ForeignKey, Text, JSON, Boolean, TIMESTAMP, func
 from sqlalchemy.orm import relationship
 from app.db.base import Base
 
 
 class QuizQuestion(Base):
     __tablename__ = "quiz_questions"
-    id = Column(BigInteger, primary_key=True)
-    lesson_id = Column(BigInteger, ForeignKey("lessons.id", ondelete="CASCADE"))
+    id = Column(Integer, primary_key=True)
+    lesson_id = Column(Integer, ForeignKey("lessons.id", ondelete="CASCADE"))
     question = Column(Text, nullable=False)
     options = Column(JSON, nullable=False) # list of options
     correct_option = Column(Integer, nullable=False)
@@ -15,9 +15,9 @@ class QuizQuestion(Base):
 
 class UserQuizAttempt(Base):
     __tablename__ = "user_quiz_attempts"
-    id = Column(BigInteger, primary_key=True)
-    user_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"))
-    lesson_id = Column(BigInteger, ForeignKey("lessons.id", ondelete="CASCADE"))
+    id = Column(Integer, primary_key=True)
+    user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"))
+    lesson_id = Column(Integer, ForeignKey("lessons.id", ondelete="CASCADE"))
     score = Column(Integer, nullable=False)
     passed = Column(Boolean, nullable=False)
     attempted_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
