+++ app/models/user.py
@@ -1,10 +1,10 @@
-from sqlalchemy import Column, BigInteger, String, TIMESTAMP, func
+from sqlalchemy import Column, Integer, String, TIMESTAMP, func
 from sqlalchemy.orm import relationship
 from app.db.base import Base
 
 class User(Base):
     __tablename__ = "users"
-    id = Column(BigInteger, primary_key=True)
+    id = Column(Integer, primary_key=True)
     email = Column(String, nullable=False, unique=True, index=True)
     display_name = Column(String, nullable=False)
     password_hash = Column(String, nullable=False)
